update application to decimate surface
