+++ CMakeLists.txt
@@ -37,8 +37,8 @@ include(${VTK_USE_FILE})
 add_executable(maskToSurfaceConstructor MaskToSurfaceConstructor.cxx)
 target_link_libraries(maskToSurfaceConstructor ${statismo_LIBRARIES} ${ITK_LIBRARIES} ${VTK_LIBRARIES})
 
-add_executable(reduceSurface ReduceSurface.cxx)
-target_link_libraries(reduceSurface ${statismo_LIBRARIES} ${ITK_LIBRARIES} ${VTK_LIBRARIES})
+add_executable(decimateSurface DecimateSurface.cxx)
+target_link_libraries(decimateSurface ${statismo_LIBRARIES} ${ITK_LIBRARIES} ${VTK_LIBRARIES})
 
 add_executable(referenceShapeConstructor ReferenceShapeConstructor.cxx)
 target_link_libraries(referenceShapeConstructor ${statismo_LIBRARIES} ${ITK_LIBRARIES})
